fix: All compoonents added in the single component folder
