+++ src/components/Disclaimer/index.js
